Got the CartItem with the correct url and clean up the ORdersPage and got the Profile using the same columnDefs
